Make main/master the default branch
This proposes we simplify our workflow to make `main` the default branch.
`master` will be renamed to `main`. Feature branches will be branched from
`main` and then merged directly to `main` without going to staging/`develop`
first.

The staging site is used ad-hoc for significant changes but is otherwise unused.